impl node_types for disk graph
