+++ raphtory-graphql/Cargo.toml
@@ -13,6 +13,7 @@ readme.workspace = true
 homepage.workspace = true
 
 [dependencies]
+raphtory-api = { path = "../raphtory-api", version = "0.8.1" }
 raphtory = { path = "../raphtory", version = "0.8.1", features = ['vectors', 'search', "io"] }
 bincode = { workspace = true }
 base64 = { workspace = true }
