+++ raphtory/src/core/utils/errors.rs
@@ -1,4 +1,5 @@
-use crate::core::{utils::time::error::ParseTimeError, ArcStr, Prop, PropType};
+use crate::core::{utils::time::error::ParseTimeError, Prop, PropType};
+use raphtory_api::core::storage::arc_str::ArcStr;
 #[cfg(feature = "search")]
 use tantivy;
 #[cfg(feature = "search")]
