+++ raphtory/src/python/graph/edge.rs
@@ -5,7 +5,7 @@
 //! edge as it existed at a particular point in time, or as it existed over a particular time range.
 //!
 use crate::{
-    core::{utils::errors::GraphError, ArcStr, Direction},
+    core::{utils::errors::GraphError, Direction},
     db::{
         api::{
             properties::Properties,
@@ -22,6 +22,7 @@ use crate::{
 use chrono::{DateTime, Utc};
 use itertools::Itertools;
 use pyo3::{prelude::*, pyclass::CompareOp};
+use raphtory_api::core::storage::arc_str::ArcStr;
 use std::{
     collections::{hash_map::DefaultHasher, HashMap},
     hash::{Hash, Hasher},
