+++ raphtory/src/python/graph/properties/constant_props.rs
@@ -1,5 +1,5 @@
 use crate::{
-    core::{ArcStr, Prop},
+    core::Prop,
     db::api::properties::{
         dyn_props::DynConstProperties, internal::PropertiesOps, ConstProperties,
     },
@@ -17,6 +17,7 @@ use pyo3::{
     exceptions::{PyKeyError, PyTypeError},
     prelude::*,
 };
+use raphtory_api::core::storage::arc_str::ArcStr;
 use std::{collections::HashMap, sync::Arc};
 
 impl<P: PropertiesOps + Send + Sync + 'static> IntoPy<PyObject> for ConstProperties<P> {
