+++ raphtory/src/python/graph/views/graph_view.rs
@@ -4,7 +4,7 @@ use rayon::prelude::*;
 use std::collections::HashMap;
 
 use crate::{
-    core::{entities::nodes::node_ref::NodeRef, utils::errors::GraphError, ArcStr},
+    core::{entities::nodes::node_ref::NodeRef, utils::errors::GraphError},
     db::{
         api::{
             properties::Properties,
@@ -34,6 +34,7 @@ use crate::{
 };
 use chrono::prelude::*;
 use pyo3::{prelude::*, types::PyBytes};
+use raphtory_api::core::storage::arc_str::ArcStr;
 
 impl IntoPy<PyObject> for MaterializedGraph {
     fn into_py(self, py: Python<'_>) -> PyObject {
